+++ src/Command/DeleteAllCommand.php
@@ -26,6 +26,7 @@ use DatabaseBackup\Utility\BackupManager;
  * Command to delete all backup files.
  *
  * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-commands#delete_all
+ * @deprecated 2.13.4' The `DeleteAllCommand` is deprecated. Will be removed in a future release
  */
 class DeleteAllCommand extends Command
 {
@@ -47,6 +48,11 @@ class DeleteAllCommand extends Command
     public function execute(Arguments $args, ConsoleIo $io): void
     {
         parent::execute($args, $io);
+
+        deprecationWarning(
+            '2.13.4',
+            'The `DeleteAllCommand` is deprecated. Will be removed in a future release'
+        );
 
         $files = BackupManager::deleteAll();
         if (!$files) {