+++ tests/TestCase/Utility/BackupExportTest.php
@@ -134,6 +134,19 @@ class BackupExportTest extends TestCase
     }
 
     /**
+     * Tests for `send()` method, is deprecated
+     *
+     * @test
+     * @uses \DatabaseBackup\Utility\BackupExport::send()
+     */
+    public function testSendIsDeprecated(): void
+    {
+        $this->deprecated(function (): void {
+            $this->BackupExport->send();
+        });
+    }
+
+    /**
      * @test
      * @uses \DatabaseBackup\Utility\BackupExport::timeout()
      */
@@ -166,6 +179,7 @@ class BackupExportTest extends TestCase
         $this->assertSame('backup.sql.bz2', basename($file));
 
         //Exports with `send()`
+        Configure::write('DatabaseBackup.mailSender', 'sender@example.com');
         $recipient = 'recipient@example.com';
         $file = $this->BackupExport->filename('exportWithSend.sql')->send($recipient)->export() ?: '';
         $this->assertMailSentFrom(Configure::readOrFail('DatabaseBackup.mailSender'));